Update activity timeout to match configuration.
The current `activity_timeout` option when establishing a connection was hardcoded to 30 seconds, but sometimes we need our application to ping at a different rate. This value should be configured through the proper config file using the `ping_interval` option which takes `REVERB_APP_PING_INTERVAL` environment variable or 60 by default.

This update takes the value which is loaded onto the `Application` class and is available through the `Connection` class, making this an easy fix.