Add an activity timeout configuration option (#241)
* Update activity timeout to match configuration.

The current `activity_timeout` option when establishing a connection was hardcoded to 30 seconds, but sometimes we need our application to ping at a different rate. This value should be configured through the proper config file using the `ping_interval` option which takes `REVERB_APP_PING_INTERVAL` environment variable or 60 by default.

This update takes the value which is loaded onto the `Application` class and is available through the `Connection` class, making this an easy fix.

* Update configuration to match old hardcoded value

This change makes sure that all future applications that re-publish the configuration keep the original 30 second value instead of the 60 second default configuration. All applications that modified this value in their environment file should not be affected.

* This update introduced the `activity_timeout` configuration to use instead of `ping_interval` to separate the concerns and accomplish the same result.

* Updated test case to include `activity_timeout` in the environment definition. Using `30` as the default value as this was the original value in the package.

* set default value

* update docblock

* Update Application.php

---------

Co-authored-by: Joe Dixon <hello@joedixon.co.uk>
Co-authored-by: Taylor Otwell <taylor@laravel.com>