Update reverb.php
Add the scheme option to the server when scaling reverb. Scheme option is used in the function to build the final redisURL, but not exposed as an option here. Often the scheme is TLS on production, making the default not work in this case, and there is no documentation on it.